nas/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
faiss-cpu>=1.7.4
numpy>=1.26.0
python-multipart>=0.0.6
httpx>=0.26.0
